ci: Parallelize op-e2e
`op-e2e` can easily be run across multiple executors. Should reduce test time from 2m to roughly 30s.
