+++ .circleci/config.yml
@@ -286,11 +286,6 @@ jobs:
             gotestsum --junitfile /test-results/op-batcher.xml -- -coverpkg=github.com/ethereum-optimism/optimism/... -coverprofile=coverage.out -covermode=atomic ./...
           working_directory: op-batcher
       - run:
-          name: test op-e2e
-          command: |
-            gotestsum --junitfile /test-results/op-e2e.xml -- -coverpkg=github.com/ethereum-optimism/optimism/... -coverprofile=coverage.out -covermode=atomic ./...
-          working_directory: op-e2e
-      - run:
           name: test op-service
           command: |
             gotestsum --junitfile /test-results/op-service.xml -- -coverpkg=github.com/ethereum-optimism/optimism/... -coverprofile=coverage.out -covermode=atomic ./...
@@ -344,6 +339,42 @@ jobs:
           name: Fuzz
           command: make fuzz
           working_directory: op-node
+
+  test-op-e2e:
+    docker:
+      - image: ethereumoptimism/ci-builder:latest
+    parallelism: 2
+    steps:
+      - checkout
+      - run:
+          name: Check if we should run
+          command: |
+            shopt -s inherit_errexit
+            CHANGED=$(check-changed "op-(batcher|bindings|e2e|node|proposer|chain-ops)" || echo "TRUE")
+            if [[ "$CHANGED" = "FALSE" ]]; then
+              circleci step halt
+            fi
+      - run:
+          name: setup
+          command: |
+            go list github.com/ethereum-optimism/optimism/... | grep -vE 'proxyd|teleportr|l2geth|indexer|endpoint-monitor|batch-submitter|gas-oracle' > /coverpkg.txt
+            echo "Covered packages:"
+            sed -z 's/\n/,/g;s/,$//' /coverpkg.txt > /coverpkg.txt.tmp
+            mv /coverpkg.txt.tmp /coverpkg.txt
+            cat /coverpkg.txt
+          working_directory: op-e2e
+      - run:
+          name: test op-e2e
+          command: |
+            SPLITS=$(go list ./... | circleci tests split --split-by=timings --timings-type=classname)
+            echo "Test splits: $SPLITS"
+            gotestsum --format standard-verbose --junitfile /test-results/op-e2e.xml -- -coverpkg=$(cat /coverpkg.txt) -coverprofile=coverage.out -covermode=atomic $SPLITS
+          working_directory: op-e2e
+      - run:
+          name: upload coverage
+          command: codecov --verbose --clean --flag bedrock-go-tests
+      - store_test_results:
+          path: /test-results
 
   depcheck:
     docker:
@@ -710,6 +741,7 @@ workflows:
           requires:
             - yarn-monorepo
       - bedrock-go-tests
+      - test-op-e2e
       - fuzz-op-node
       - bedrock-markdown
       - devnet
